gralloc: always just return 0 from check_for_compression() with no checks on the 7580
In check_for_compression(), the value of the property ddk.set.afbc
was checked. We don't have it set, and the 7580 doesn't support AFBC
(Arm Frame Buffer Compression) so it can't be set to 1. If that property
is set to 0, check_for_compression() always returned 0. So, cut the
checks and simply always do that if the target SoC is the Exynos7580.

The reading of the value of ddk.set.afbc causes SELinux denials under 12,
so by doing this, the inconvenience of addressing them (properly with a
new proper property context) just for it is avoided.

Change-Id: I49399e49113c3dd0e24456e8e501f685ef64d790